+++ consensus/merge/src/test/java/org/hyperledger/besu/consensus/merge/blockcreation/MergeReorgTest.java
@@ -91,7 +91,8 @@ public class MergeReorgTest implements MergeGenesisConfigHelper {
             CompletableFuture::runAsync,
             mockPendingTransactions,
             new MiningParameters.Builder().coinbase(coinbase).build(),
-            mock(BackwardSyncContext.class));
+            mock(BackwardSyncContext.class),
+            Optional.empty());
     mergeContext.setIsPostMerge(genesisState.getBlock().getHeader().getDifficulty());
     blockchain.observeBlockAdded(
         blockAddedEvent ->